# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NexWatch Hub API
  description: >
    Custom API routes exposed by the NexWatch hub (a PocketBase application)
    under /api/custom/*, plus the unauthenticated /healthz liveness probe.
    PocketBase's own built-in routes (record CRUD under /api/collections/*,
    auth, realtime, files, etc.) are not documented here — see the
    PocketBase API reference for those.
  version: "1.0.0"
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/custom/logs:
    get:
      operationId: getLogs
      summary: Search shipped log entries
      description: >
        Returns "logs" rows (internal/agent/logs shipping journald and
        file-tailed lines to the hub), newest first. "q" is a
        case-insensitive substring match against the message text.
        "before" is an opaque cursor (the previous response's
        next_before) for "load older" pagination; live tail instead
        subscribes to PocketBase realtime on the "logs" collection
        directly, which is not a "/api/custom" route.
      tags: [logs]
      parameters:
        - name: agent_id
          in: query
          schema: { type: string }
          description: Restrict to one agent. Omit for every agent.
        - name: level
          in: query
          schema: { type: string, enum: [error, warning, info, debug] }
        - name: unit
          in: query
          schema: { type: string }
          description: Exact match against the systemd unit (journald sources only).
        - name: q
          in: query
          schema: { type: string }
          description: Case-insensitive substring match against the message.
        - name: since
          in: query
          schema: { type: integer }
          description: Unix milliseconds. Only entries at or after this time.
        - name: until
          in: query
          schema: { type: integer }
          description: Unix milliseconds. Only entries at or before this time.
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 500 }
        - name: before
          in: query
          schema: { type: string }
          description: Cursor from a previous response's next_before, for loading older entries.
      responses:
        "200":
          description: Matching log entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: "#/components/schemas/LogEntry"
                  next_before:
                    type: string
                    description: Cursor to pass as "before" for the next older page. Absent once there are no more results.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/custom/logs/units:
    get:
      operationId: getLogUnits
      summary: Distinct log units seen in the last 24h
      description: >
        Returns the distinct "unit" values across "logs" rows from the
        last 24 hours (optionally scoped to one agent), for populating a
        unit filter dropdown. A file-tailed source has no unit and is
        excluded.
      tags: [logs]
      parameters:
        - name: agent_id
          in: query
          schema: { type: string }
          description: Restrict to one agent. Omit for every agent.
      responses:
        "200":
          description: Distinct unit names, sorted alphabetically.
          content:
            application/json:
              schema:
                type: object
                properties:
                  units:
                    type: array
                    items: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        A PocketBase auth token (superuser or "users" record) passed in the
        Authorization header. PocketBase accepts it either as a bare token
        or with a "Bearer " prefix.
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ServerError:
      description: An unexpected server-side error occurred.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    LogEntry:
      type: object
      properties:
        id: { type: string }
        agent_id: { type: string }
        ts: { type: integer, description: Unix milliseconds. }
        source:
          type: string
          description: "\"journald\", or \"file:<path>\" for a tailed file."
        unit: { type: string, description: systemd unit name (journald sources only). }
        level: { type: string, enum: [error, warning, info, debug] }
        message: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
          description: Small set of extra structured attributes (e.g. journald's SYSLOG_IDENTIFIER).